Tokenizes the input bibtex file.
It does catch some errors however there is a bunch of errors like
missing a ',' which it does not catch yet. The string searching is
slow. Switch Case Used as a hashmap and does not gurantee input validity
yet.

A weird bug arises when the 'current' reaches the end of file. It
generates an error. It does not however affect the file processing.