Fixing a Scanning Bug
Due to how the scanner works, we need to move through all fields. Nothing is actually tokenized so
in some cases where the 'value' contains a ',' or '}' the toknizer failed. We changed this so that
now the string is completely scanned even though it is not stored. Fortunately this has a added benifit
having a opurtunity to report "unknown key" warnings